backend_py/requirements.txt
# FastAPI 核心依赖
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy>=2.0.36
nanoid==2.0.0

# 认证相关 (新增)
python-jose[cryptography]>=3.3.0
pydantic[email]>=2.0.0
email-validator>=2.1.0

# LLM 相关依赖
requests==2.31.0
openai==2.6.1 

# 文档解析
chardet==5.2.0
pypdf==3.17.4
python-docx==1.1.0
beautifulsoup4==4.12.3
lxml==6.0.2

# 测试依赖
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.23.3

# 代码质量
pylint==3.0.3
black==23.12.1
tox==4.11.4